move buildNewServer to its own file
